End of section 10

+++ Application/Activities/Delete.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -9,11 +10,11 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command,Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -21,11 +22,18 @@ namespace Application.Activities
             {
                 _context = context;
             }
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity=await _context.Activities.FindAsync(request.Id);
+                if(activity==null)
+                {
+                    return null;
+                }
                 _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
+                var result =await _context.SaveChangesAsync()>0;
+                if(!result)
+                    return Result<Unit>.Failure("Failed to delete the activity");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
